refactor(shared): remove .gitkeep files, redundant ref, and unify CQRS folder
