+++ src/Ecommerce.Catalog/Ecommerce.Catalog.Application/Categories/CreateCategory/CreateCategoryCommand.cs
@@ -1,4 +1,4 @@
-using Ecommerce.Shared.Application.Commands;
+using Ecommerce.Shared.Application.CQRS;
 
 namespace Ecommerce.Catalog.Application.Categories.CreateCategory;
