+++ src/Ecommerce.Catalog/Ecommerce.Catalog.Application/Ecommerce.Catalog.Application.csproj
@@ -2,7 +2,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Ecommerce.Catalog.Domain\Ecommerce.Catalog.Domain.csproj" />
-    <ProjectReference Include="..\..\Ecommerce.Shared\Ecommerce.Shared.Domain\Ecommerce.Shared.Domain.csproj" />
     <ProjectReference Include="..\..\Ecommerce.Shared\Ecommerce.Shared.Application\Ecommerce.Shared.Application.csproj" />
   </ItemGroup>
