+++ src/Ecommerce.Shared/Ecommerce.Shared.Application/CQRS/IQuery.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace Ecommerce.Shared.Application.Queries;
+namespace Ecommerce.Shared.Application.CQRS;
 
 public interface IQuery<out TResult> : IRequest<TResult>
 {
